requirements.txt
flask
pyyaml
attrdict
psycopg2
parse
httpx
gunicorn
argon2-cffi
Flask-WTF
Flask-SQLAlchemy
Flask-Migrate
Flask-Script